WIP. Adding new tests. Refactoring.
