+++ mm2src/mm2_tests.rs
@@ -62,6 +62,10 @@ use electrums::*;
 mod lightning_tests;
 
 #[cfg(all(test, not(target_arch = "wasm32")))]
+#[path = "mm2_tests/orderbook_sync_tests.rs"]
+mod orderbook_sync_tests;
+
+#[cfg(all(test, not(target_arch = "wasm32")))]
 #[path = "mm2_tests/lp_bot_tests.rs"]
 mod lp_bot_tests;
 
@@ -261,208 +265,6 @@ fn local_start() -> LocalStart { local_start_impl }
 
 #[cfg(target_arch = "wasm32")]
 fn local_start() -> LocalStart { wasm_start_impl }
-
-/// https://github.com/artemii235/SuperNET/issues/241
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-fn alice_can_see_the_active_order_after_connection() {
-    let coins = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","rpcport":11608,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}}
-    ]);
-
-    // start bob and immediately place the order
-    let mm_bob = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": "bob passphrase",
-            "coins": coins,
-            "rpc_password": "pass",
-            "i_am_seed": true,
-        }),
-        "pass".into(),
-        local_start!("bob"),
-    )
-    .unwrap();
-    let (_bob_dump_log, _bob_dump_dashboard) = mm_bob.mm_dump();
-    log!({"Bob log path: {}", mm_bob.log_path.display()});
-    // Enable coins on Bob side. Print the replies in case we need the "address".
-    log!({ "enable_coins (bob): {:?}", block_on(enable_coins_rick_morty_electrum(&mm_bob)) });
-    // issue sell request on Bob side by setting base/rel price
-    log!("Issue bob sell request");
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "MORTY",
-        "price": 0.9,
-        "volume": "0.9",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-    // Bob orderbook must show the new order
-    log!("Get RICK/MORTY orderbook on Bob side");
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let bob_orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("Bob orderbook "[bob_orderbook]);
-    assert!(!bob_orderbook.asks.is_empty(), "Bob RICK/MORTY asks are empty");
-    assert_eq!(BigDecimal::from_str("0.9").unwrap(), bob_orderbook.asks[0].max_volume);
-
-    // start eve and immediately place the order
-    let mm_eve = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": "eve passphrase",
-            "coins": coins,
-            "rpc_password": "pass",
-            "seednodes": [fomat!((mm_bob.ip))],
-        }),
-        "pass".into(),
-        local_start!("bob"),
-    )
-    .unwrap();
-    let (_eve_dump_log, _eve_dump_dashboard) = mm_eve.mm_dump();
-    log!({ "Eve log path: {}", mm_eve.log_path.display() });
-    // Enable coins on Eve side. Print the replies in case we need the "address".
-    log!({ "enable_coins (eve): {:?}", block_on(enable_coins_rick_morty_electrum(&mm_eve)) });
-    // issue sell request on Eve side by setting base/rel price
-    log!("Issue eve sell request");
-    let rc = block_on(mm_eve.rpc(json! ({
-        "userpass": mm_eve.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "MORTY",
-        "price": "1",
-        "volume": "0.9",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-    // issue sell request on Eve side by setting base/rel price
-    log!("Issue eve sell request");
-    let rc = block_on(mm_eve.rpc(json! ({
-        "userpass": mm_eve.userpass,
-        "method": "setprice",
-        "base": "MORTY",
-        "rel": "RICK",
-        "price": "1",
-        "volume": "0.9",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    log!("Get RICK/MORTY orderbook on Eve side");
-    let rc = block_on(mm_eve.rpc(json! ({
-        "userpass": mm_eve.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let eve_orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("Eve orderbook "[eve_orderbook]);
-    assert_eq!(
-        eve_orderbook.asks.len(),
-        2,
-        "Eve RICK/MORTY orderbook must have exactly 2 asks"
-    );
-    assert_eq!(
-        eve_orderbook.bids.len(),
-        1,
-        "Eve RICK/MORTY orderbook must have exactly 1 bid"
-    );
-
-    log!("Give Bob 2 seconds to import Eve order");
-    thread::sleep(Duration::from_secs(2));
-    log!("Get RICK/MORTY orderbook on Bob side");
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let bob_orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("Bob orderbook "[bob_orderbook]);
-    assert_eq!(
-        bob_orderbook.asks.len(),
-        2,
-        "Bob RICK/MORTY orderbook must have exactly 2 asks"
-    );
-    assert_eq!(
-        bob_orderbook.bids.len(),
-        1,
-        "Bob RICK/MORTY orderbook must have exactly 1 bid"
-    );
-
-    let mm_alice = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("ALICE_TRADE_IP") .ok(),
-            "rpcip": env::var ("ALICE_TRADE_IP") .ok(),
-            "passphrase": "alice passphrase",
-            "coins": coins,
-            "seednodes": [fomat!((mm_bob.ip))],
-            "rpc_password": "pass",
-        }),
-        "pass".into(),
-        local_start!("alice"),
-    )
-    .unwrap();
-
-    let (_alice_dump_log, _alice_dump_dashboard) = mm_alice.mm_dump();
-    log!({ "Alice log path: {}", mm_alice.log_path.display() });
-
-    // Enable coins on Alice side. Print the replies in case we need the "address".
-    log!({ "enable_coins (alice): {:?}", block_on(enable_coins_rick_morty_electrum(&mm_alice)) });
-
-    log!("Get RICK/MORTY orderbook on Alice side");
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let alice_orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("Alice orderbook "[alice_orderbook]);
-    assert_eq!(
-        alice_orderbook.asks.len(),
-        2,
-        "Alice RICK/MORTY orderbook must have exactly 2 asks"
-    );
-    assert_eq!(
-        alice_orderbook.bids.len(),
-        1,
-        "Alice RICK/MORTY orderbook must have exactly 1 bid"
-    );
-
-    block_on(mm_bob.stop()).unwrap();
-    block_on(mm_alice.stop()).unwrap();
-    block_on(mm_eve.stop()).unwrap();
-}
 
 /// https://github.com/KomodoPlatform/atomicDEX-API/issues/886#issuecomment-812489844
 #[test]
@@ -1126,13 +928,6 @@ async fn trade_base_rel_electrum(
 
     Timer::sleep(1.).await;
 
-    // Both Alice and Bob might try to bind on the "0.0.0.0:47773" DHT port in this test
-    // (because the local "127.0.0.*:47773" addresses aren't that useful for DHT).
-    // We want to give Bob a headstart in acquiring the port,
-    // because Alice will then be able to directly reach it (thanks to "seednode").
-    // Direct communication is not required in this test, but it's nice to have.
-    // wait_log_re! (mm_bob, 9., "preferred port");
-
     let mut mm_alice = MarketMakerIt::start_async(
         json! ({
             "gui": "nogui",
@@ -1176,9 +971,9 @@ async fn trade_base_rel_electrum(
 
         std::fs::copy("./mm2src/coins/for_tests/ZOMBIE_CACHE.db", alice_zombie_cache_path).unwrap();
 
-        let zombie_bob = enable_native(&mm_bob, "ZOMBIE", &["http://195.201.0.6:8565"]).await;
+        let zombie_bob = enable_native(&mm_bob, "ZOMBIE", &[]).await;
         log!("enable ZOMBIE bob "[zombie_bob]);
-        let zombie_alice = enable_native(&mm_alice, "ZOMBIE", &["http://195.201.0.6:8565"]).await;
+        let zombie_alice = enable_native(&mm_alice, "ZOMBIE", &[]).await;
         log!("enable ZOMBIE alice "[zombie_alice]);
     }
     // Enable coins on Bob side. Print the replies in case we need the address.
@@ -2604,301 +2399,6 @@ fn test_own_orders_should_not_be_removed_from_orderbook() {
     block_on(mm_bob.stop()).unwrap();
 }
 
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-// https://github.com/KomodoPlatform/atomicDEX-API/issues/511
-fn test_all_orders_per_pair_per_node_must_be_displayed_in_orderbook() {
-    let coins = json!([
-        {"coin":"RICK","asset":"RICK","protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","protocol":{"type":"UTXO"}},
-    ]);
-
-    let mm = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": "bob passphrase",
-            "coins": coins,
-            "rpc_password": "pass",
-            "i_am_seed": true,
-        }),
-        "pass".into(),
-        match var("LOCAL_THREAD_MM") {
-            Ok(ref e) if e == "bob" => Some(local_start()),
-            _ => None,
-        },
-    )
-    .unwrap();
-    let (_dump_log, _dump_dashboard) = mm.mm_dump();
-    log!({"Log path: {}", mm.log_path.display()});
-    block_on(enable_electrum(&mm, "RICK", false, &[
-        "electrum3.cipig.net:10017",
-        "electrum2.cipig.net:10017",
-        "electrum1.cipig.net:10017",
-    ]));
-    block_on(enable_electrum(&mm, "MORTY", false, &[
-        "electrum3.cipig.net:10018",
-        "electrum2.cipig.net:10018",
-        "electrum1.cipig.net:10018",
-    ]));
-
-    // set 2 orders with different prices
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "MORTY",
-        "price": 0.9,
-        "volume": "0.9",
-        "cancel_previous": false,
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "MORTY",
-        "price": 1,
-        "volume": "0.9",
-        "cancel_previous": false,
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    thread::sleep(Duration::from_secs(2));
-
-    log!("Get RICK/MORTY orderbook");
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let orderbook: Json = json::from_str(&rc.1).unwrap();
-    log!("orderbook "[orderbook]);
-    let asks = orderbook["asks"].as_array().unwrap();
-    assert_eq!(asks.len(), 2, "RICK/MORTY orderbook must have exactly 2 asks");
-}
-
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-// https://github.com/KomodoPlatform/atomicDEX-API/issues/859
-fn orderbook_extended_data() {
-    let coins = json!([
-        {"coin":"RICK","asset":"RICK","protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","protocol":{"type":"UTXO"}},
-    ]);
-
-    let mm = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": "bob passphrase",
-            "coins": coins,
-            "rpc_password": "pass",
-            "i_am_seed": true,
-        }),
-        "pass".into(),
-        match var("LOCAL_THREAD_MM") {
-            Ok(ref e) if e == "bob" => Some(local_start()),
-            _ => None,
-        },
-    )
-    .unwrap();
-    let (_dump_log, _dump_dashboard) = &mm.mm_dump();
-    log!({"Log path: {}", mm.log_path.display()});
-    block_on(enable_electrum(&mm, "RICK", false, &[
-        "electrum3.cipig.net:10017",
-        "electrum2.cipig.net:10017",
-        "electrum1.cipig.net:10017",
-    ]));
-    block_on(enable_electrum(&mm, "MORTY", false, &[
-        "electrum3.cipig.net:10018",
-        "electrum2.cipig.net:10018",
-        "electrum1.cipig.net:10018",
-    ]));
-
-    let bob_orders = &[
-        // (base, rel, price, volume)
-        ("RICK", "MORTY", "0.9", "0.9"),
-        ("RICK", "MORTY", "0.8", "0.9"),
-        ("RICK", "MORTY", "0.7", "0.9"),
-        ("MORTY", "RICK", "0.8", "0.9"),
-        ("MORTY", "RICK", "1", "0.9"),
-    ];
-
-    for (base, rel, price, volume) in bob_orders {
-        let rc = block_on(mm.rpc(json!({
-            "userpass": mm.userpass,
-            "method": "setprice",
-            "base": base,
-            "rel": rel,
-            "price": price,
-            "volume": volume,
-            "cancel_previous": false,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-    }
-
-    thread::sleep(Duration::from_secs(1));
-    log!("Get RICK/MORTY orderbook");
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("orderbook "[rc.1]);
-    let expected_total_asks_base_vol = MmNumber::from("2.7");
-    assert_eq!(expected_total_asks_base_vol.to_decimal(), orderbook.total_asks_base_vol);
-
-    let expected_total_bids_base_vol = MmNumber::from("1.62");
-    assert_eq!(expected_total_bids_base_vol.to_decimal(), orderbook.total_bids_base_vol);
-
-    let expected_total_asks_rel_vol = MmNumber::from("2.16");
-    assert_eq!(expected_total_asks_rel_vol.to_decimal(), orderbook.total_asks_rel_vol);
-
-    let expected_total_bids_rel_vol = MmNumber::from("1.8");
-    assert_eq!(expected_total_bids_rel_vol.to_decimal(), orderbook.total_bids_rel_vol);
-
-    fn check_price_and_vol_aggr(
-        order: &OrderbookEntryAggregate,
-        price: &'static str,
-        base_aggr: &'static str,
-        rel_aggr: &'static str,
-    ) {
-        let price = MmNumber::from(price);
-        assert_eq!(price.to_decimal(), order.price);
-
-        let base_aggr = MmNumber::from(base_aggr);
-        assert_eq!(base_aggr.to_decimal(), order.base_max_volume_aggr);
-
-        let rel_aggr = MmNumber::from(rel_aggr);
-        assert_eq!(rel_aggr.to_decimal(), order.rel_max_volume_aggr);
-    }
-
-    check_price_and_vol_aggr(&orderbook.asks[0], "0.9", "2.7", "2.16");
-    check_price_and_vol_aggr(&orderbook.asks[1], "0.8", "1.8", "1.35");
-    check_price_and_vol_aggr(&orderbook.asks[2], "0.7", "0.9", "0.63");
-
-    check_price_and_vol_aggr(&orderbook.bids[0], "1.25", "0.72", "0.9");
-    check_price_and_vol_aggr(&orderbook.bids[1], "1", "1.62", "1.8");
-}
-
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-fn orderbook_should_display_base_rel_volumes() {
-    let coins = json!([
-        {"coin":"RICK","asset":"RICK","protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","protocol":{"type":"UTXO"}},
-    ]);
-
-    let mm = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": "bob passphrase",
-            "coins": coins,
-            "rpc_password": "pass",
-            "i_am_seed": true,
-        }),
-        "pass".into(),
-        match var("LOCAL_THREAD_MM") {
-            Ok(ref e) if e == "bob" => Some(local_start()),
-            _ => None,
-        },
-    )
-    .unwrap();
-    let (_dump_log, _dump_dashboard) = &mm.mm_dump();
-    log!({"Log path: {}", mm.log_path.display()});
-    block_on(enable_electrum(&mm, "RICK", false, &[
-        "electrum3.cipig.net:10017",
-        "electrum2.cipig.net:10017",
-        "electrum1.cipig.net:10017",
-    ]));
-    block_on(enable_electrum(&mm, "MORTY", false, &[
-        "electrum3.cipig.net:10018",
-        "electrum2.cipig.net:10018",
-        "electrum1.cipig.net:10018",
-    ]));
-
-    let price = BigRational::new(2.into(), 1.into());
-    let volume = BigRational::new(1.into(), 1.into());
-
-    // create order with rational amount and price
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "MORTY",
-        "price": price,
-        "volume": volume,
-        "cancel_previous": false,
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    thread::sleep(Duration::from_secs(1));
-    log!("Get RICK/MORTY orderbook");
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("orderbook "[orderbook]);
-    assert_eq!(orderbook.asks.len(), 1, "RICK/MORTY orderbook must have exactly 1 ask");
-    let min_volume = BigRational::new(1.into(), 10000.into());
-    assert_eq!(volume, orderbook.asks[0].base_max_volume_rat);
-    assert_eq!(min_volume, orderbook.asks[0].base_min_volume_rat);
-
-    assert_eq!(&volume * &price, orderbook.asks[0].rel_max_volume_rat);
-    assert_eq!(&min_volume * &price, orderbook.asks[0].rel_min_volume_rat);
-
-    log!("Get MORTY/RICK orderbook");
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "orderbook",
-        "base": "MORTY",
-        "rel": "RICK",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("orderbook "[orderbook]);
-    assert_eq!(orderbook.bids.len(), 1, "MORTY/RICK orderbook must have exactly 1 bid");
-    let min_volume = BigRational::new(1.into(), 10000.into());
-    assert_eq!(volume, orderbook.bids[0].rel_max_volume_rat);
-    assert_eq!(min_volume, orderbook.bids[0].rel_min_volume_rat);
-
-    assert_eq!(&volume * &price, orderbook.bids[0].base_max_volume_rat);
-    assert_eq!(&min_volume * &price, orderbook.bids[0].base_min_volume_rat);
-}
-
 #[cfg(not(target_arch = "wasm32"))]
 fn check_priv_key(mm: &MarketMakerIt, coin: &str, expected_priv_key: &str) {
     let rc = block_on(mm.rpc(json! ({
@@ -3148,216 +2648,6 @@ fn setprice_buy_sell_too_low_volume() {
     check_too_low_volume_order_creation_fails(&mm, "MORTY", "ETH");
     check_too_low_volume_order_creation_fails(&mm, "ETH", "MORTY");
     check_too_low_volume_order_creation_fails(&mm, "JST", "MORTY");
-}
-
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-// https://github.com/KomodoPlatform/atomicDEX-API/issues/473
-fn setprice_min_volume_should_be_displayed_in_orderbook() {
-    let bob_passphrase = get_passphrase(&".env.seed", "BOB_PASSPHRASE").unwrap();
-
-    let coins = json! ([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","rpcport":11608,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"ETH","name":"ethereum","protocol":{"type":"ETH"}},
-        {"coin":"JST","name":"jst","protocol":{"type":"ERC20","protocol_data":{"platform":"ETH","contract_address":"0x2b294F029Fde858b2c62184e8390591755521d8E"}}}
-    ]);
-
-    let mm_bob = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": bob_passphrase,
-            "coins": coins,
-            "rpc_password": "pass",
-            "i_am_seed": true,
-        }),
-        "pass".into(),
-        match var("LOCAL_THREAD_MM") {
-            Ok(ref e) if e == "bob" => Some(local_start()),
-            _ => None,
-        },
-    )
-    .unwrap();
-
-    let (_dump_log, _dump_dashboard) = mm_bob.mm_dump();
-    log!({"Bob log path: {}", mm_bob.log_path.display()});
-
-    let mm_alice = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "dht": "on",  // Enable DHT without delay.
-            "myipaddr": env::var ("ALICE_TRADE_IP") .ok(),
-            "rpcip": env::var ("ALICE_TRADE_IP") .ok(),
-            "passphrase": "alice passphrase",
-            "coins": coins,
-            "seednodes": [fomat!((mm_bob.ip))],
-            "rpc_password": "pass",
-        }),
-        "pass".into(),
-        match var("LOCAL_THREAD_MM") {
-            Ok(ref e) if e == "alice" => Some(local_start()),
-            _ => None,
-        },
-    )
-    .unwrap();
-
-    let (_alice_dump_log, _alice_dump_dashboard) = mm_alice.mm_dump();
-    log!({"Alice log path: {}", mm_alice.log_path.display()});
-
-    log! ({"enable_coins (bob): {:?}", block_on (enable_coins_eth_electrum (&mm_bob, &["http://195.201.0.6:8565"]))});
-    log! ({"enable_coins (alice): {:?}", block_on (enable_coins_eth_electrum (&mm_alice, &["http://195.201.0.6:8565"]))});
-
-    // issue orderbook call on Alice side to trigger subscription to a topic
-    block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "orderbook",
-        "base": "ETH",
-        "rel": "JST",
-    })))
-    .unwrap();
-
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "setprice",
-        "base": "ETH",
-        "rel": "JST",
-        "price": "1",
-        "volume": "10",
-        "min_volume": "1",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    thread::sleep(Duration::from_secs(2));
-    log!("Get ETH/JST orderbook on Bob side");
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "orderbook",
-        "base": "ETH",
-        "rel": "JST",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let orderbook: Json = json::from_str(&rc.1).unwrap();
-    log!("orderbook "[orderbook]);
-    let asks = orderbook["asks"].as_array().unwrap();
-    assert_eq!(asks.len(), 1, "Bob ETH/JST orderbook must have exactly 1 ask");
-
-    let min_volume = asks[0]["min_volume"].as_str().unwrap();
-    assert_eq!(min_volume, "1", "Bob ETH/JST ask must display correct min_volume");
-
-    log!("Get ETH/JST orderbook on Alice side");
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "orderbook",
-        "base": "ETH",
-        "rel": "JST",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let orderbook: Json = json::from_str(&rc.1).unwrap();
-    log!("orderbook "[orderbook]);
-    let asks = orderbook["asks"].as_array().unwrap();
-    assert_eq!(asks.len(), 1, "Alice ETH/JST orderbook must have exactly 1 ask");
-
-    let min_volume = asks[0]["min_volume"].as_str().unwrap();
-    assert_eq!(min_volume, "1", "Alice ETH/JST ask must display correct min_volume");
-}
-
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-// https://github.com/KomodoPlatform/atomicDEX-API/issues/670
-fn orderbook_should_work_without_coins_activation() {
-    let bob_passphrase = get_passphrase(&".env.seed", "BOB_PASSPHRASE").unwrap();
-
-    let coins = json! ([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","rpcport":11608,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"ETH","name":"ethereum","protocol":{"type":"ETH"}},
-        {"coin":"JST","name":"jst","protocol":{"type":"ERC20","protocol_data":{"platform":"ETH","contract_address":"0x2b294F029Fde858b2c62184e8390591755521d8E"}}}
-    ]);
-
-    let mm_bob = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": bob_passphrase,
-            "coins": coins,
-            "rpc_password": "pass",
-            "i_am_seed": true,
-        }),
-        "pass".into(),
-        match var("LOCAL_THREAD_MM") {
-            Ok(ref e) if e == "bob" => Some(local_start()),
-            _ => None,
-        },
-    )
-    .unwrap();
-
-    let (_dump_log, _dump_dashboard) = mm_bob.mm_dump();
-    log!({"Bob log path: {}", mm_bob.log_path.display()});
-
-    let mm_alice = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "dht": "on",  // Enable DHT without delay.
-            "myipaddr": env::var ("ALICE_TRADE_IP") .ok(),
-            "rpcip": env::var ("ALICE_TRADE_IP") .ok(),
-            "passphrase": "alice passphrase",
-            "coins": coins,
-            "seednodes": [fomat!((mm_bob.ip))],
-            "rpc_password": "pass",
-        }),
-        "pass".into(),
-        match var("LOCAL_THREAD_MM") {
-            Ok(ref e) if e == "alice" => Some(local_start()),
-            _ => None,
-        },
-    )
-    .unwrap();
-
-    let (_alice_dump_log, _alice_dump_dashboard) = mm_alice.mm_dump();
-    log!({"Alice log path: {}", mm_alice.log_path.display()});
-
-    log! ({"enable_coins (bob): {:?}", block_on (enable_coins_eth_electrum (&mm_bob, &["http://195.201.0.6:8565"]))});
-
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "setprice",
-        "base": "ETH",
-        "rel": "JST",
-        "price": "1",
-        "volume": "10",
-        "min_volume": "1",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    log!("Get ETH/JST orderbook on Alice side");
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "orderbook",
-        "base": "ETH",
-        "rel": "JST",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let orderbook: Json = json::from_str(&rc.1).unwrap();
-    log!("orderbook "[orderbook]);
-    let asks = orderbook["asks"].as_array().unwrap();
-    assert_eq!(asks.len(), 1, "Alice ETH/JST orderbook must have exactly 1 ask");
 }
 
 #[test]
@@ -7658,343 +6948,6 @@ fn test_best_orders_address_and_confirmations() {
     block_on(mm_alice.stop()).unwrap();
 }
 
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-fn test_orderbook_segwit() {
-    let bob_passphrase = get_passphrase(&".env.seed", "BOB_PASSPHRASE").unwrap();
-
-    let bob_coins_config = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"tBTC","name":"tbitcoin","fname":"tBitcoin","rpcport":18332,"pubtype":111,"p2shtype":196,"wiftype":239,"segwit":true,"bech32_hrp":"tb","txfee":0,"estimate_fee_mode":"ECONOMICAL","mm2":1,"required_confirmations":0,"protocol":{"type":"UTXO"},"address_format":{"format":"segwit"}}
-    ]);
-
-    let alice_coins_config = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"tBTC","name":"tbitcoin","fname":"tBitcoin","rpcport":18332,"pubtype":111,"p2shtype":196,"wiftype":239,"segwit":true,"bech32_hrp":"tb","txfee":0,"estimate_fee_mode":"ECONOMICAL","mm2":1,"required_confirmations":0,"protocol":{"type":"UTXO"}}
-    ]);
-
-    let mut mm_bob = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": bob_passphrase,
-            "coins": bob_coins_config,
-            "rpc_password": "pass",
-            "i_am_seed": true,
-        }),
-        "pass".into(),
-        local_start!("bob"),
-    )
-    .unwrap();
-    let (_bob_dump_log, _bob_dump_dashboard) = mm_bob.mm_dump();
-    log!({"Bob log path: {}", mm_bob.log_path.display()});
-
-    // Enable coins on Bob side. Print the replies in case we need the "address".
-    let electrum = block_on(mm_bob.rpc(json!({
-        "userpass": "pass",
-        "method": "electrum",
-        "coin": "tBTC",
-        "servers": [{"url":"electrum1.cipig.net:10068"},{"url":"electrum2.cipig.net:10068"},{"url":"electrum3.cipig.net:10068"}],
-        "address_format":{"format":"segwit"},
-        "mm2": 1,
-    }))).unwrap();
-    assert_eq!(
-        electrum.0,
-        StatusCode::OK,
-        "RPC «electrum» failed with {} {}",
-        electrum.0,
-        electrum.1
-    );
-    log!({ "enable tBTC: {:?}", electrum });
-    let enable_tbtc_res: EnableElectrumResponse = json::from_str(&electrum.1).unwrap();
-    let tbtc_segwit_address = enable_tbtc_res.address;
-
-    let electrum = block_on(mm_bob.rpc(json!({
-        "userpass": "pass",
-        "method": "electrum",
-        "coin": "RICK",
-        "servers": [{"url":"electrum1.cipig.net:10017"},{"url":"electrum2.cipig.net:10017"},{"url":"electrum3.cipig.net:10017"}],
-        "mm2": 1,
-    }))).unwrap();
-    assert_eq!(
-        electrum.0,
-        StatusCode::OK,
-        "RPC «electrum» failed with {} {}",
-        electrum.0,
-        electrum.1
-    );
-    log!({ "enable RICK: {:?}", electrum });
-    let enable_rick_res: Json = json::from_str(&electrum.1).unwrap();
-    let rick_address = enable_rick_res["address"].as_str().unwrap();
-
-    // issue sell request on Bob side by setting base/rel price
-    log!("Issue bob sell requests");
-
-    let bob_orders = [
-        // (base, rel, price, volume, min_volume)
-        ("tBTC", "RICK", "0.7", "0.0002", Some("0.00015")),
-        ("RICK", "tBTC", "0.7", "0.0002", Some("0.00015")),
-    ];
-    for (base, rel, price, volume, min_volume) in bob_orders.iter() {
-        let rc = block_on(mm_bob.rpc(json! ({
-            "userpass": mm_bob.userpass,
-            "method": "setprice",
-            "base": base,
-            "rel": rel,
-            "price": price,
-            "volume": volume,
-            "min_volume": min_volume.unwrap_or("0.00777"),
-            "cancel_previous": false,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-    }
-
-    let mm_alice = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("ALICE_TRADE_IP") .ok(),
-            "rpcip": env::var ("ALICE_TRADE_IP") .ok(),
-            "passphrase": "alice passphrase",
-            "coins": alice_coins_config,
-            "seednodes": [fomat!((mm_bob.ip))],
-            "rpc_password": "pass",
-        }),
-        "pass".into(),
-        local_start!("alice"),
-    )
-    .unwrap();
-
-    let (_alice_dump_log, _alice_dump_dashboard) = mm_alice.mm_dump();
-    log!({ "Alice log path: {}", mm_alice.log_path.display() });
-
-    block_on(mm_bob.wait_for_log(22., |log| {
-        log.contains("DEBUG Handling IncludedTorelaysMesh message for peer")
-    }))
-    .unwrap();
-
-    // checking orderbook on alice side
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "orderbook",
-        "base": "tBTC",
-        "rel": "RICK",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-    let response: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    assert_eq!(response.asks[0].address, tbtc_segwit_address);
-    assert_eq!(response.bids[0].address, rick_address);
-
-    block_on(mm_bob.stop()).unwrap();
-    block_on(mm_alice.stop()).unwrap();
-}
-
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-fn alice_can_see_the_active_order_after_orderbook_sync_segwit() {
-    let bob_passphrase = get_passphrase(&".env.seed", "BOB_PASSPHRASE").unwrap();
-
-    let bob_coins_config = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"tBTC","name":"tbitcoin","fname":"tBitcoin","rpcport":18332,"pubtype":111,"p2shtype":196,"wiftype":239,"segwit":true,"bech32_hrp":"tb","txfee":0,"estimate_fee_mode":"ECONOMICAL","mm2":1,"required_confirmations":0,"protocol":{"type":"UTXO"},"address_format":{"format":"segwit"}}
-    ]);
-
-    let alice_coins_config = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"tBTC","name":"tbitcoin","fname":"tBitcoin","rpcport":18332,"pubtype":111,"p2shtype":196,"wiftype":239,"segwit":true,"bech32_hrp":"tb","txfee":0,"estimate_fee_mode":"ECONOMICAL","mm2":1,"required_confirmations":0,"protocol":{"type":"UTXO"},"address_format":{"format":"segwit"}}
-    ]);
-
-    let mut mm_bob = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("BOB_TRADE_IP") .ok(),
-            "rpcip": env::var ("BOB_TRADE_IP") .ok(),
-            "canbind": env::var ("BOB_TRADE_PORT") .ok().map (|s| s.parse::<i64>().unwrap()),
-            "passphrase": bob_passphrase,
-            "coins": bob_coins_config,
-            "rpc_password": "pass",
-            "i_am_seed": true,
-        }),
-        "pass".into(),
-        local_start!("bob"),
-    )
-    .unwrap();
-    let (_bob_dump_log, _bob_dump_dashboard) = mm_bob.mm_dump();
-    log!({"Bob log path: {}", mm_bob.log_path.display()});
-
-    // Enable coins on Bob side. Print the replies in case we need the "address".
-    let electrum = block_on(mm_bob.rpc(json!({
-        "userpass": "pass",
-        "method": "electrum",
-        "coin": "tBTC",
-        "servers": [{"url":"electrum1.cipig.net:10068"},{"url":"electrum2.cipig.net:10068"},{"url":"electrum3.cipig.net:10068"}],
-        "address_format":{"format":"segwit"},
-        "mm2": 1,
-    }))).unwrap();
-    assert_eq!(
-        electrum.0,
-        StatusCode::OK,
-        "RPC «electrum» failed with {} {}",
-        electrum.0,
-        electrum.1
-    );
-    log!({ "enable tBTC: {:?}", electrum });
-    let enable_tbtc_res: EnableElectrumResponse = json::from_str(&electrum.1).unwrap();
-    let tbtc_segwit_address = enable_tbtc_res.address;
-
-    let electrum = block_on(mm_bob.rpc(json!({
-        "userpass": "pass",
-        "method": "electrum",
-        "coin": "RICK",
-        "servers": [{"url":"electrum1.cipig.net:10017"},{"url":"electrum2.cipig.net:10017"},{"url":"electrum3.cipig.net:10017"}],
-        "mm2": 1,
-    }))).unwrap();
-    assert_eq!(
-        electrum.0,
-        StatusCode::OK,
-        "RPC «electrum» failed with {} {}",
-        electrum.0,
-        electrum.1
-    );
-    log!({ "enable RICK: {:?}", electrum });
-    let enable_rick_res: Json = json::from_str(&electrum.1).unwrap();
-    let rick_address = enable_rick_res["address"].as_str().unwrap();
-
-    // issue sell request on Bob side by setting base/rel price
-    log!("Issue bob sell requests");
-
-    let bob_orders = [
-        // (base, rel, price, volume, min_volume)
-        ("tBTC", "RICK", "0.7", "0.0002", Some("0.00015")),
-        ("RICK", "tBTC", "0.7", "0.0002", Some("0.00015")),
-    ];
-    for (base, rel, price, volume, min_volume) in bob_orders.iter() {
-        let rc = block_on(mm_bob.rpc(json! ({
-            "userpass": mm_bob.userpass,
-            "method": "setprice",
-            "base": base,
-            "rel": rel,
-            "price": price,
-            "volume": volume,
-            "min_volume": min_volume.unwrap_or("0.00777"),
-            "cancel_previous": false,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-    }
-
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "mmrpc": "2.0",
-        "method": "get_public_key",
-        "params": {},
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!get_public_key: {}", rc.1);
-    let get_public_key_res: RpcV2Response<GetPublicKeyResult> = serde_json::from_str(&rc.1).unwrap();
-    let bob_pubkey = get_public_key_res.result.public_key;
-
-    let mut mm_alice = MarketMakerIt::start(
-        json! ({
-            "gui": "nogui",
-            "netid": 9998,
-            "myipaddr": env::var ("ALICE_TRADE_IP") .ok(),
-            "rpcip": env::var ("ALICE_TRADE_IP") .ok(),
-            "passphrase": "alice passphrase",
-            "coins": alice_coins_config,
-            "seednodes": [fomat!((mm_bob.ip))],
-            "rpc_password": "pass",
-        }),
-        "pass".into(),
-        local_start!("alice"),
-    )
-    .unwrap();
-
-    let (_alice_dump_log, _alice_dump_dashboard) = mm_alice.mm_dump();
-    log!({ "Alice log path: {}", mm_alice.log_path.display() });
-
-    block_on(mm_bob.wait_for_log(22., |log| {
-        log.contains("DEBUG Handling IncludedTorelaysMesh message for peer")
-    }))
-    .unwrap();
-
-    let electrum = block_on(mm_alice.rpc(json!({
-        "userpass": "pass",
-        "method": "electrum",
-        "coin": "tBTC",
-        "servers": [{"url":"electrum1.cipig.net:10068"},{"url":"electrum2.cipig.net:10068"},{"url":"electrum3.cipig.net:10068"}],
-        "address_format":{"format":"segwit"},
-        "mm2": 1,
-    }))).unwrap();
-    assert_eq!(
-        electrum.0,
-        StatusCode::OK,
-        "RPC «electrum» failed with {} {}",
-        electrum.0,
-        electrum.1
-    );
-    log!({ "enable Alice tBTC: {:?}", electrum });
-
-    let electrum = block_on(mm_alice.rpc(json!({
-        "userpass": "pass",
-        "method": "electrum",
-        "coin": "RICK",
-        "servers": [{"url":"electrum1.cipig.net:10017"},{"url":"electrum2.cipig.net:10017"},{"url":"electrum3.cipig.net:10017"}],
-        "mm2": 1,
-    }))).unwrap();
-    assert_eq!(
-        electrum.0,
-        StatusCode::OK,
-        "RPC «electrum» failed with {} {}",
-        electrum.0,
-        electrum.1
-    );
-    log!({ "enable Alice RICK: {:?}", electrum });
-
-    // setting the price will trigger Alice's subscription to the orderbook topic
-    // but won't request the actual orderbook
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "tBTC",
-        "price": "1",
-        "volume": "0.1",
-        "cancel_previous": false,
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    block_on(
-        mm_alice.wait_for_log((MIN_ORDER_KEEP_ALIVE_INTERVAL * 2) as f64, |log| {
-            log.contains(&format!("Inserting order OrderbookItem {{ pubkey: \"{}\"", bob_pubkey))
-        }),
-    )
-    .unwrap();
-
-    // checking orderbook on alice side
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "orderbook",
-        "base": "tBTC",
-        "rel": "RICK",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-    let response: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    assert_eq!(response.asks[0].address, tbtc_segwit_address);
-    assert_eq!(response.bids[0].address, rick_address);
-
-    block_on(mm_bob.stop()).unwrap();
-    block_on(mm_alice.stop()).unwrap();
-}
-
 #[cfg(not(target_arch = "wasm32"))]
 fn request_and_check_orderbook_depth(mm_alice: &MarketMakerIt) {
     let rc = block_on(mm_alice.rpc(json! ({
@@ -8218,283 +7171,4 @@ fn test_get_public_key() {
         v.result.public_key,
         "022cd3021a2197361fb70b862c412bc8e44cff6951fa1de45ceabfdd9b4c520420"
     )
-}
-
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-fn test_get_orderbook_with_same_orderbook_ticker() {
-    let coins = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"protocol":{"type":"UTXO"}},
-        {"coin":"RICK-Utxo","asset":"RICK","orderbook_ticker":"RICK","rpcport":8923,"txversion":4,"protocol":{"type":"UTXO"}},
-        // just a random contract address
-        {"coin":"RICK-ERC20","orderbook_ticker":"RICK","decimals": 18,"protocol":{"type":"ERC20","protocol_data":{"platform":"ETH","contract_address":"0x7Fc66500c84A76Ad7e9c93437bFc5Ac33E2DDaE9"}}},
-    ]);
-
-    let mm = MarketMakerIt::start(
-        json!({
-            "gui": "nogui",
-            "netid": 9998,
-            "passphrase": "bob passphrase",
-            "rpc_password": "password",
-            "coins": coins,
-            "i_am_seed": true,
-        }),
-        "password".into(),
-        None,
-    )
-    .unwrap();
-    let (_dump_log, _dump_dashboard) = mm.mm_dump();
-    log!({"Log path: {}", mm.log_path.display()});
-
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "RICK-Utxo",
-    })))
-    .unwrap();
-    assert!(
-        rc.0.is_server_error(),
-        "orderbook succeed but should have failed {}",
-        rc.1
-    );
-
-    let rc = block_on(mm.rpc(json! ({
-        "userpass": mm.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "RICK-ERC20",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook {}", rc.1);
-}
-
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-fn test_conf_settings_in_orderbook() {
-    let coins = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"required_confirmations":10,"requires_notarization":true,"protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","rpcport":11608,"txversion":4,"overwintered":1,"required_confirmations":5,"requires_notarization":false,"protocol":{"type":"UTXO"}},
-    ]);
-
-    let mm_bob = MarketMakerIt::start(
-        json!({
-            "gui": "nogui",
-            "netid": 9998,
-            "passphrase": "bob passphrase",
-            "rpc_password": "password",
-            "coins": coins,
-            "i_am_seed": true,
-        }),
-        "password".into(),
-        None,
-    )
-    .unwrap();
-    let (_dump_log, _dump_dashboard) = mm_bob.mm_dump();
-    log!({"Log path: {}", mm_bob.log_path.display()});
-
-    log! ({"enable_coins (bob): {:?}", block_on (enable_coins_rick_morty_electrum(&mm_bob))});
-
-    log!("Issue set_price request for RICK/MORTY on Bob side");
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "MORTY",
-        "price": 0.9,
-        "volume": "0.9",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    log!("Issue set_price request for MORTY/RICK on Bob side");
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "setprice",
-        "base": "MORTY",
-        "rel": "RICK",
-        "price": 0.9,
-        "volume": "0.9",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    let mm_alice = MarketMakerIt::start(
-        json!({
-            "gui": "nogui",
-            "netid": 9998,
-            "passphrase": "alice passphrase",
-            "rpc_password": "password",
-            "coins": coins,
-            "seednodes": [fomat!((mm_bob.ip))],
-        }),
-        "password".into(),
-        None,
-    )
-    .unwrap();
-    let (_dump_log, _dump_dashboard) = mm_alice.mm_dump();
-    log!({"Log path: {}", mm_alice.log_path.display()});
-
-    log! ({"enable_coins (alice): {:?}", block_on (enable_coins_rick_morty_electrum(&mm_alice))});
-
-    log!("Get RICK/MORTY orderbook on Alice side");
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let alice_orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("Alice orderbook "[alice_orderbook]);
-
-    assert_eq!(
-        alice_orderbook.asks.len(),
-        1,
-        "Alice RICK/MORTY orderbook must have exactly 1 ask"
-    );
-    assert_eq!(alice_orderbook.asks[0].base_confs, 10);
-    assert_eq!(alice_orderbook.asks[0].base_nota, true);
-    assert_eq!(alice_orderbook.asks[0].rel_confs, 5);
-    assert_eq!(alice_orderbook.asks[0].rel_nota, false);
-
-    assert_eq!(
-        alice_orderbook.bids.len(),
-        1,
-        "Alice RICK/MORTY orderbook must have exactly 1 bid"
-    );
-    assert_eq!(alice_orderbook.bids[0].base_confs, 10);
-    assert_eq!(alice_orderbook.bids[0].base_nota, true);
-    assert_eq!(alice_orderbook.bids[0].rel_confs, 5);
-    assert_eq!(alice_orderbook.bids[0].rel_nota, false);
-
-    block_on(mm_bob.stop()).unwrap();
-    block_on(mm_alice.stop()).unwrap();
-}
-
-#[test]
-#[cfg(not(target_arch = "wasm32"))]
-fn alice_can_see_confs_in_orderbook_after_sync() {
-    let bob_coins = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"required_confirmations":10,"requires_notarization":true,"protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","rpcport":11608,"txversion":4,"overwintered":1,"required_confirmations":5,"requires_notarization":false,"protocol":{"type":"UTXO"}},
-    ]);
-
-    let mm_bob = MarketMakerIt::start(
-        json!({
-            "gui": "nogui",
-            "netid": 9998,
-            "passphrase": "bob passphrase",
-            "rpc_password": "password",
-            "coins": bob_coins,
-            "i_am_seed": true,
-        }),
-        "password".into(),
-        None,
-    )
-    .unwrap();
-    // let (_dump_log, _dump_dashboard) = mm_bob.mm_dump();
-    log!({"Bob log path: {}", mm_bob.log_path.display()});
-
-    log! ({"enable_coins (bob): {:?}", block_on (enable_coins_rick_morty_electrum(&mm_bob))});
-
-    log!("Issue sell request on Bob side");
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "MORTY",
-        "price": 0.9,
-        "volume": "0.9",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    let rc = block_on(mm_bob.rpc(json! ({
-        "userpass": mm_bob.userpass,
-        "mmrpc": "2.0",
-        "method": "get_public_key",
-        "params": {},
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!get_public_key: {}", rc.1);
-    let get_public_key_res: RpcV2Response<GetPublicKeyResult> = serde_json::from_str(&rc.1).unwrap();
-    let bob_pubkey = get_public_key_res.result.public_key;
-
-    // Alice coins don't have required_confirmations and requires_notarization set
-    let alice_coins = json!([
-        {"coin":"RICK","asset":"RICK","rpcport":8923,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-        {"coin":"MORTY","asset":"MORTY","rpcport":11608,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
-    ]);
-
-    let mut mm_alice = MarketMakerIt::start(
-        json!({
-            "gui": "nogui",
-            "netid": 9998,
-            "passphrase": "alice passphrase",
-            "rpc_password": "password",
-            "coins": alice_coins,
-            "seednodes": [fomat!((mm_bob.ip))],
-        }),
-        "password".into(),
-        None,
-    )
-    .unwrap();
-    let (_dump_log, _dump_dashboard) = mm_alice.mm_dump();
-    log!({"Alice log path: {}", mm_alice.log_path.display()});
-
-    log! ({"enable_coins (alice): {:?}", block_on (enable_coins_rick_morty_electrum(&mm_alice))});
-
-    // setting the price will trigger Alice's subscription to the orderbook topic
-    // but won't request the actual orderbook
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "setprice",
-        "base": "RICK",
-        "rel": "MORTY",
-        "price": "1",
-        "volume": "0.1",
-        "cancel_previous": false,
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-    block_on(
-        mm_alice.wait_for_log((MIN_ORDER_KEEP_ALIVE_INTERVAL * 2) as f64, |log| {
-            log.contains(&format!("Inserting order OrderbookItem {{ pubkey: \"{}\"", bob_pubkey))
-        }),
-    )
-    .unwrap();
-
-    log!("Get RICK/MORTY orderbook on Alice side");
-    let rc = block_on(mm_alice.rpc(json! ({
-        "userpass": mm_alice.userpass,
-        "method": "orderbook",
-        "base": "RICK",
-        "rel": "MORTY",
-    })))
-    .unwrap();
-    assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-
-    let alice_orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-    log!("Alice orderbook "[alice_orderbook]);
-    assert_eq!(
-        alice_orderbook.asks.len(),
-        2,
-        "Alice RICK/MORTY orderbook must have exactly 2 ask"
-    );
-    let bob_order_in_orderbook = alice_orderbook
-        .asks
-        .iter()
-        .find(|entry| entry.pubkey == bob_pubkey)
-        .unwrap();
-    assert_eq!(bob_order_in_orderbook.base_confs, 10);
-    assert_eq!(bob_order_in_orderbook.base_nota, true);
-    assert_eq!(bob_order_in_orderbook.rel_confs, 5);
-    assert_eq!(bob_order_in_orderbook.rel_nota, false);
-
-    block_on(mm_bob.stop()).unwrap();
-    block_on(mm_alice.stop()).unwrap();
 }
